feat(departement): add DepartementNotFoundException, update DepartementService and DepartementRepository to handle not found scenarios, and enhance DepartementController for error handling

+++ api/src/main/java/org/acme/departement/domain/port/in/DepartementRest.java
@@ -8,5 +8,5 @@ public interface DepartementRest {
 
     List<Departement> listDepartement();
 
-    Departement findByDepartementId(int id);
+    Departement findByDepartementId(Long id);
 }
